+++ src/swap/constants.ts
@@ -11,3 +11,15 @@ export const UNCAUGHT_SWAP_ERROR_CODE = 'UNCAUGHT_SWAP_ERROR';
 export const UNIVERSALROUTER_CONTRACT_ADDRESS =
   '0x3fC91A3afd70395Cd496C647d5a6CC9D4B2b7FAD';
 export const USER_REJECTED_ERROR_CODE = 'USER_REJECTED';
+export enum SwapMessage {
+  BALANCE_ERROR = 'Error fetching token balance',
+  CONFIRM_IN_WALLET = 'Confirm in wallet',
+  FETCHING_QUOTE = 'Fetching quote...',
+  FETCHING_BALANCE = 'Fetching balance...',
+  INCOMPLETE_FIELD = 'Complete the fields to continue',
+  INSUFFICIENT_BALANCE = 'Insufficient balance',
+  LOW_LIQUIDITY = 'Liquidity too low for the token',
+  SWAP_IN_PROGRESS = 'Swap in progress...',
+  TOO_MANY_REQUESTS = 'Too many requests. Please try again later.',
+  USER_REJECTED = 'User rejected the transaction',
+}
